backend/requirements.txt
boto3==1.34.0
python-jose==3.3.0
flask==2.3.3
werkzeug==2.3.7